CNDA scripts finalized and CHPC transfer scripts

+++ scripts/MRI/CNDA/CNDA_get_new_sessions.sh
@@ -19,15 +19,25 @@ bash /data/perlman/moochie/github/LCBDtools/scripts/MRI/CNDA/CNDA_get_sessions.s
 
 # for any new sessions (diff between old and new)
 # run python emailer to remind RAs to go enter usability
-for new_session in `comm -23 <(sort "/data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions_new.txt") <(sort "/data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions.txt")`; do
-	
-	/usr/bin/python3.7 /data/perlman/moochie/github/LCBDtools/scripts/MRI/CNDA/email_new_CNDA_session.py $new_session
 
-done
-
+# if the new file exists
 if [[ -f /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions_new.txt ]]; then
+	
+	# and if it has more lines than the original one
+	if [[ $(wc -l < /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions_new.txt) -ge $(wc -l < /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions.txt) ]]; then
 
-	rm /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions.txt
-	mv /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions_new.txt /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions.txt
+			
+		for new_session in `comm -23 <(sort "/data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions_new.txt") <(sort "/data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions.txt")`
+		do
+			/usr/bin/python3.7 /data/perlman/moochie/github/LCBDtools/scripts/MRI/CNDA/email_new_CNDA_session.py $new_session
+		done
+
+		rm /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions.txt
+		mv /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions_new.txt /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions.txt
+
+	# if it doesn't, something went wrong (probably fail to reach CNDA) and just skip this time around by removing new file
+	else
+		rm /data/perlman/moochie/study_data/CARE/CNDA_downloads/CNDA_sessions_new.txt
+	fi
 
 fi
